+++ src/components/ModalSignIn/ModalSignIn.js
@@ -32,9 +32,8 @@ const ModalSignIn = (props) => {
       password: "",
     },
     validationSchema: loginUserValidationSchema,
-    onSubmit: async (values, { setSubmitting, resetForm }) => {
+    onSubmit: async (values, { setSubmitting, resetForm, setErrors }) => {
       try {
-        console.log(values, "async");
         const response = await axiosInterceptors.post(REQ.LOGIN_USER, {
           login_email: values.email,
           login_pass: values.password,
@@ -43,7 +42,13 @@ const ModalSignIn = (props) => {
         toast.success("User login successful!");
         resetForm();
       } catch (error) {
-        toast.error("Error logging in. Please try again.");
+        const errorMessage =
+          error?.response?.data?.error ||
+          error?.response?.data?.message ||
+          "Error logging in. Please try again.";
+
+        toast.error(errorMessage);
+        setErrors({ apiError: errorMessage });
       } finally {
         setSubmitting(false);
       }
